garden trees, flowers, plant added

+++ week-04/day-2/Garden2/src/com/greenfoxacademy/Plants.java
@@ -3,14 +3,17 @@ package com.greenfoxacademy;
 public class Plants {
 String color;
 Boolean thirsty;
-int thirstLevel;
+double thirstLevel;
 
   public Plants(String color) {
     this.color = color;
   }
 
   public Plants() {
+
   }
+
+
 
 
 }
